Getting backin login to work on fastapi

+++ api/routers/JobPostForm.py
@@ -10,38 +10,35 @@ from queries.JobForm_queries import (
     JobPostFormOut2
 )
 from queries.accounts import AccountRepo
-from RoleChecker import RoleChecker
+from authenticator import authenticator
 
 router = APIRouter()
 
-checker = RoleChecker("Employer")
-
 # creating new job form
-@router.post('/create_form/{account_id}', tags=["JobForm"], response_model=JobPostFormOut2)
+@router.post('/create_form/', tags=["JobForm"], response_model=JobPostFormOut2)
 def create_job_form(
     new_form: JobPostFormIn,
-    account_id: int,
     repo: JobFormRepository = Depends(),
     repo1: AccountRepo = Depends(),
-    checked_role: bool = Depends(checker),
+    account: dict = Depends(authenticator.get_current_account_data)
 ):
     credentials_exception = HTTPException(
         status_code=status.HTTP_403_FORBIDDEN,
         detail="You do not have access to this as employee. TURN BACK NOW!",
     )
-    if checked_role:
-        not_final = repo.create(new_form, account_id).dict()
-        not_final["account_id"] = repo1.get(account_id).dict()
+    if account["role"] == "Employer":
+        not_final = repo.create(new_form, account["id"]).dict()
+        not_final["account_id"] = account
         return not_final
     raise credentials_exception
 
 
 @router.get("/get_all_form", tags=["JobForm"], response_model=Union[List[JobPostForm], Error])
-def get_all_job_form(repo: JobFormRepository = Depends()):
+def get_all_job_form(repo: JobFormRepository = Depends(), account: dict = Depends(authenticator.get_current_account_data)):
     return repo.get_all()
 
 @router.get("/get_form/{form_id}", tags=["JobForm"], response_model=Union[JobPostFormOut, Error])
-def get_one_job_form(form_id: int, response: Response, repo: JobFormRepository = Depends()) -> JobPostFormOut:
+def get_one_job_form(form_id: int, response: Response, repo: JobFormRepository = Depends(), account: dict = Depends(authenticator.get_current_account_data)) -> JobPostFormOut:
     FormDetail = repo.get_one(form_id)
     if FormDetail is None:
         response.status_code = 404
@@ -52,20 +49,18 @@ def Update_Job_Form(
     form_id: int,
     UpdatedJobForm: JobPostFormIn,
     repo: JobFormRepository = Depends(),
-    checked_role: bool = Depends(checker),
+    account: dict = Depends(authenticator.get_current_account_data)
 ) -> Union[JobPostFormOut, Error]:
     credentials_exception = HTTPException(
         status_code=status.HTTP_403_FORBIDDEN,
         detail="You are employee. Only the employer can edit",
     )
-    if checked_role:
+    if account["role"] == "Employer":
         return repo.update(form_id, UpdatedJobForm)
-        # not_final = repo.create(new_form, account_id).dict()  ####
-        # not_final["account_id"] = repo1.get(account_id).dict()
-        # return not_final
     raise credentials_exception
 
 
 @router.delete("/delete_job_form/{id}", tags=["JobForm"], response_model=bool)
-def Delete_Job_Form(form_id: int, repo: JobFormRepository = Depends()) -> bool:
-    return repo.delete(form_id)
+def Delete_Job_Form(form_id: int, repo: JobFormRepository = Depends(), account: dict = Depends(authenticator.get_current_account_data)) -> bool:
+    if account["role"] == "Employer":
+        return repo.delete(form_id)
